requirements.txt
Django>=2.2,<3.0
djoser
jsonschema>=3.2,<3.3
psycopg2>=2.8,<2.9
djangorestframework>=3.10.0,<3.11.0
djangorestframework_simplejwt
djoser
django-environ
django-cors-headers
graphene-django
django-graphql-jwt
django-celery-beat
celery>=5.0.0
redis
pillow
